+++ testClasses/RectangleMethodsTest.java
@@ -11,18 +11,18 @@ public class RectangleMethodsTest extends TestCase{
 
 	@Test
 	public void testMoveNorth(){
-		Rectangle2D.Double move = constructRectangle();
-		Rectangle2D.Double compare = constructRectangle();
+		Rectangle2D.Float move = constructRectangle();
+		Rectangle2D.Float compare = constructRectangle();
 		float length = (float) 0.5;
-		compare.y = 100.0; 
+		compare.y = 100.0f; 
 		move = RectangleMethods.move(move, length, Direction.NORTH);
 		assertEquals(compare, move);
 	}
 	
 	@Test
 	public void testMoveSouth(){
-		Rectangle2D.Double move = constructRectangle();
-		Rectangle2D.Double compare = constructRectangle();
+		Rectangle2D.Float move = constructRectangle();
+		Rectangle2D.Float compare = constructRectangle();
 		float length = (float) 0.5;
 		compare.y = 0; 
 		move = RectangleMethods.move(move, length, Direction.SOUTH);
@@ -31,18 +31,18 @@ public class RectangleMethodsTest extends TestCase{
 	
 	@Test
 	public void testMoveEast(){
-		Rectangle2D.Double move = constructRectangle();
-		Rectangle2D.Double compare = constructRectangle();
+		Rectangle2D.Float move = constructRectangle();
+		Rectangle2D.Float compare = constructRectangle();
 		float length = (float) 0.5;
-		compare.x = 100.0; 
+		compare.x = 100.0f; 
 		move = RectangleMethods.move(move, length, Direction.EAST);
 		assertEquals(compare, move);
 	}
 	
 	@Test
 	public void testMoveWest(){
-		Rectangle2D.Double move = constructRectangle();
-		Rectangle2D.Double compare = constructRectangle();
+		Rectangle2D.Float move = constructRectangle();
+		Rectangle2D.Float compare = constructRectangle();
 		float length = (float) 0.5;
 		compare.x = 0; 
 		move = RectangleMethods.move(move, length, Direction.WEST);
@@ -51,8 +51,8 @@ public class RectangleMethodsTest extends TestCase{
 	
 	@Test
 	public void testFixRatioByInnerRectangle(){
-		Rectangle2D.Double inner = constructRectangle();
-		Rectangle2D.Double outer = constructRectangle();
+		Rectangle2D.Float inner = constructRectangle();
+		Rectangle2D.Float outer = constructRectangle();
 		outer.width *= 1.5; outer.height *= 2.5;
 		RectangleMethods.fixRatioByInnerRectangle(inner, outer);
 		float inner_ratio = (float) (inner.width/inner.height);
@@ -62,8 +62,8 @@ public class RectangleMethodsTest extends TestCase{
 	
 	@Test
 	public void testFixRatioByOuterRectangle(){
-		Rectangle2D.Double inner = constructRectangle();
-		Rectangle2D.Double outer = constructRectangle();
+		Rectangle2D.Float inner = constructRectangle();
+		Rectangle2D.Float outer = constructRectangle();
 		outer.width *= 1.5; outer.height *= 2.5;
 		RectangleMethods.fixRatioByOuterRectangle(inner, outer);
 		float inner_ratio = (float) (inner.width/inner.height);
@@ -73,16 +73,16 @@ public class RectangleMethodsTest extends TestCase{
 	
 	@Test
 	public void testPoint2DToRectangle(){
-		Point2D.Double a = new Point2D.Double(50, 150);
-		Point2D.Double b = new Point2D.Double(150, 50);
-		Rectangle2D.Double compare = RectangleMethods.point2DToRectangle(a, b);
+		Point2D.Float a = new Point2D.Float(50, 150);
+		Point2D.Float b = new Point2D.Float(150, 50);
+		Rectangle2D.Float compare = RectangleMethods.point2DToRectangle(a, b);
 		assertEquals(compare, constructRectangle());
 	}
 	
 	@Test
 	public void testZoomRectangle(){
-		Rectangle2D.Double zoom = RectangleMethods.zoomRectangle((float)0.2, true, constructRectangle());
-		Rectangle2D.Double compare = constructRectangle();
+		Rectangle2D.Float zoom = RectangleMethods.zoomRectangle((float)0.2, true, constructRectangle());
+		Rectangle2D.Float compare = constructRectangle();
 		compare.width = 60;
 		compare.height = 60;
 		compare.x = 70;
@@ -110,7 +110,7 @@ public class RectangleMethodsTest extends TestCase{
 	 * 
 	 * @return A rectangle with x,y at 50, 50 and width and height of 100
 	 */
-	private Rectangle2D.Double constructRectangle(){
-		return new Rectangle2D.Double(50, 50, 100, 100);
+	private Rectangle2D.Float constructRectangle(){
+		return new Rectangle2D.Float(50, 50, 100, 100);
 	}
 }
